Added a way to kill stalled jobs / pods

+++ kubernetes/run_data_generation.py
@@ -141,7 +141,7 @@ def run_generation(params):
 
             time.sleep(wait_time_sec)
 
-            if(k % 2 == 0): # every 2 min including the 1st min, check progress
+            if(k % 5 == 0): # every 2 min including the 1st min, check progress
 
                 config.load_kube_config()
                 v1 = client.CoreV1Api()
@@ -150,27 +150,29 @@ def run_generation(params):
                 #pod_list = v1.list_namespaced_pod(namespace = params["namespace"])
                 #pod_names = [item.metadata.name for item in pod_list.items]
 
-                pod_statuses = [item.status.phase for item in pod_list.items]
-                pod_phases = [1 for ele in pod_statuses if(ele == "Succeeded")]
-
-                pod_times_min = []
-                for item in pod_list.items:
-                    then = item.status.start_time
-                    now = datetime.datetime.now(tzutc())
-                    pod_times_min.append((now - then).total_seconds() / 60)
-
-                hit_list = [i for i, time in enumerate(pod_times_min) if(time >= params["kill_time_min"])]
- 
                 if(k == 0):
                     print()
 
-                for i, job_name in enumerate(current_group):
-                    if(i in hit_list):
-                        print("Removing job: %s" % job_name)
-                        subprocess.run(["kubectl", "delete", "job", job_name])
-                        current_group.pop(i)
-                        pod_statuses.pop(i)
-                        pod_phases.pop(i)
+                if(k != 0):
+
+                    pod_statuses = [item.status.phase for item in pod_list.items]
+                    pod_phases = [1 for ele in pod_statuses if(ele == "Succeeded")]
+
+                    pod_times_min = []
+                    for item in pod_list.items:
+                        then = item.status.start_time
+                        now = datetime.datetime.now(tzutc())
+                        pod_times_min.append((now - then).total_seconds() / 60)
+
+                    hit_list = [i for i, time in enumerate(pod_times_min) if(time >= params["kill_time_min"])]
+     
+                    for i, job_name in enumerate(current_group):
+                        if(i in hit_list):
+                            print("Removing job: %s" % job_name)
+                            subprocess.run(["kubectl", "delete", "job", job_name])
+                            current_group.pop(i)
+                            pod_statuses.pop(i)
+                            pod_phases.pop(i)
 
                 #print("Progressing group %s: Elapsed Time = %s minutes, Completion = %s / %s" % (parallel_id, (wait_time_sec * (k + 1)) / 60, sum(pod_phases), params["num_parallel_ops"]))
